// src/appointments/types/appointment.types.ts
export enum AppointmentStatus {
  PENDING_PAYMENT = 'PENDING_PAYMENT',
  SCHEDULED = 'SCHEDULED',
  IN_PROGRESS = 'IN_PROGRESS',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
  NO_SHOW = 'NO_SHOW',
  INTERRUPTED = 'INTERRUPTED',
}

export enum AppointmentType {
  NEW_PROBLEM = 'NEW_PROBLEM',
  FOLLOW_UP = 'FOLLOW_UP',
}

export enum TerminationReason {
  // CANCELLED
  CUSTOMER_CANCELLED_EARLY = 'CUSTOMER_CANCELLED_EARLY',
  CUSTOMER_CANCELLED_LATE = 'CUSTOMER_CANCELLED_LATE',
  PAYMENT_TIMEOUT = 'PAYMENT_TIMEOUT',
  SYSTEM_CANCELLED = 'SYSTEM_CANCELLED',

  //  NO_SHOW
  CUSTOMER_NO_SHOW = 'CUSTOMER_NO_SHOW',
  DOCTOR_NO_SHOW = 'DOCTOR_NO_SHOW',

  //  INTERRUPTED
  TECHNICAL_ISSUE_CUSTOMER = 'TECHNICAL_ISSUE_CUSTOMER',
  TECHNICAL_ISSUE_DOCTOR = 'TECHNICAL_ISSUE_DOCTOR',
  TECHNICAL_ISSUE_PLATFORM = 'TECHNICAL_ISSUE_PLATFORM',
  EMERGENCY_LEAVE_CUSTOMER = 'EMERGENCY_LEAVE_CUSTOMER',
  EMERGENCY_LEAVE_DOCTOR = 'EMERGENCY_LEAVE_DOCTOR',
}